// src/interfaces/home.ts
export interface Guarantee {
  name: string;
  slug: string;
  icon: string;
}
export interface Client {
  image: string;
  width: number;
  height: number;
}

export interface Innovation {
  icon: string;
  slug: string;
  name: string;
  text: string;
}

export interface Responsibility {
  title: string;
  text: string;
  image: string;
  logos: string[];
}